requirements.txt
Django==1.5.1
South==0.7.6
argparse==1.2.1
distribute==0.6.24
django-crispy-forms==1.2.7
django-recaptcha==0.0.6
django-tinymce==1.5.1
ipdb==0.7
ipython==0.13.2
wsgiref==0.1.2
